`src/cpu.rs`: Combine other `cpu.rs`'s into one for simplicity.
